eval/rag_eval_builder/requirements.txt
fpdf2>=2.7.0
openpyxl>=3.1.0
faker>=20.0.0
tqdm>=4.65.0
typer>=0.9.0
